audioflinger: Throttle output if no active tracks
Throttle output for the duration from when there are
no active tracks to the time output is put to
standby due to inactivity. This is done to save some
power during this duration. A typical usecase affected
by this optimization is key press on the dialer
keypad or keyboard with a delay of a few hundred
milliseconds between keys.

Change-Id: Ibf6a3ada6094b6d077d4b8634ee2137942702a55
CRs-Fixed: 2088743