audio: ensure effect chain with specific session id is unique
It's possible that tracks with the same session id running on various
playback outputs, which causes effect chain being created on the same
session twice. As a result, the same effect engine will be released
twice as the same context is reused.

Output that has effect chain with same session id is more preferable.

Change-Id: I5a777949a532739c139bfc026630ebe8477016ed
CRs-Fixed: 2232451